requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
channels>=4.0.0
channels-redis>=4.1.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
Pillow>=10.0.0
django-filter>=23.0
pytest>=7.4.0
pytest-django>=4.7.0
pytest-asyncio>=0.21.0
coverage>=7.3.0
